OBJ loader basically working. Normals not looking right but I think that's unrelated to the OBJ loader
Former-commit-id: 27a9e3be95c3caa44b8375b0f2724752c0a1ff7e

+++ OpenGL_Template/ShapeContainer.cpp
@@ -249,7 +249,7 @@ ShapeData::indicesType ShapeContainer::depthSort(glm::vec3 s_cam_location)
 		auto numIndices = 0;
 		for (GLuint j = 0; j < i.second; ++j) 
 		{
-			numIndices += m_shapes.at(j).numVertices();
+			numIndices += (int)m_shapes.at(j).numVertices();
 		}
 		auto shape = std::move(m_shapes.at(i.second));
 		for (auto it = shape.indx_begin(); it != shape.indx_end(); ++it)
